Made small fixes in credits_router and implemented services of creditsRouter in front-end

+++ ai_personal_content_app/lib/core/common/models/user_credits_and_quotas_model.dart
@@ -10,7 +10,7 @@ abstract class UserCreditsAndQuotasModel with _$UserCreditsAndQuotasModel {
     required String id,
     required String userId,
     required int remainingAdsQuotaForToday,
-    required int remainingCredits,
+    required double remainingCredits,
     int? lastAdWatchTime,
     int? lastCreditUseTime,
   }) = _UserCreditsAndQuotasModel;
